# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Chee's Life"
  description: 'Use "Accept-Language" as lang Header, support `tc` or `en`'
  version: 1.0.0
servers:
  -
    url: 'https://chees-api.dev.com'
paths:
...
  /api/auth/user/update-password:
    post:
      summary: 'Update Password'
      operationId: updatePassword
      description: 'Update password'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses: {  }
      tags:
        - 'CMS API'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  description: ''
                  example: recusandae
                newPassword:
                  type: string
                  description: ''
                  example: recusandae
              required:
                - currentPassword
                - newPassword
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []